Use a universal abstraction for constrained collection types
This could be used for different types of constraints to implement
types such as `Map`, `BitSet`, `SortedSet`. Computing the result type
works similarly to `CanBuild` but the eligible implicit scopes and the
relationship of different implicits are much simpler. In particular,
there is never a `From` type and there are only two choices for the
`To` type: the desired constrained type (which does not need to be a
type constructor) and an unconstrained fallback type constructor.

The current implementation shows that inheritance works (with plenty of
`@uncheckedVariance` tricks) if the constraint is the same. We add an
`Ordering` constraint in `SortedSet` and extend that to `TreeSet`. If
the element type has an `Ordering` poly-transforms produce another
`TreeSet` (or `SortedSet`, depending on the original type), otherwise a
`Set`.

Unfortunately, it does not look like it will be possible to compose
multiple constraints, which would be needed for types like `TreeMap`
(combining the `Tuple2` constraint of `Map` with the `Ordering`
constraint of `Sorted`).