WIP: Bounded factories
Adding an upper bound to element types allows `FromIterable` and
`IterableFactory` to be used for building constrained collection types.
In cases where the constraint is only a type proof (e.g. `BitSet`,
`String`, `Map`) they are on equal footing with unconstrained
collections.

As a proof of concept, I added a `BitSet` class. If you have an
`Iterable[Int]` you can now call `.to(BitSet)` with no overloading or
runtime overhead (like passing an additional implicit parameter).

Sort-of-problem: We need pretend unary type constructors for all
collection types, so instead of a `BitSet` you really get a
`BitSet with Iterable[T]`.

This still doesn’t help with the actual implementation of constrained
collection types (only with their companion objects).

Next step: Implicit lookup of companion objects and extending it to
constraints with a runtime component.